# pkg/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Forkspacer API
  version: 1.0.0
  description: API for managing workspaces and modules in Forkspacer
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /workspace/connection/kubeconfig/:
    post:
      summary: Create a kubeconfig secret
      operationId: createKubeconfigSecret
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - kubeconfig
              properties:
                name:
                  type: string
                  description: DNS 1123 subdomain name
                  pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$'
                  maxLength: 253
                kubeconfig:
                  type: string
                  format: binary
                  description: Kubeconfig file (max 10 MB)
      responses:
        '201':
          description: Kubeconfig secret created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      success:
                        allOf:
                          - $ref: '#/components/schemas/JSONSuccessResponse'
                          - type: object
                            properties:
                              code:
                                type: string
                                enum: [created]
                              data:
                                $ref: '#/components/schemas/KubeconfigSecretResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/FormDataTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      summary: Delete a kubeconfig secret
      operationId: deleteKubeconfigSecret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteKubeconfigSecretRequest'
      responses:
        '204':
          description: Kubeconfig secret deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /workspace/connection/kubeconfig/list:
    get:
      summary: List kubeconfig secrets
      operationId: listKubeconfigSecrets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListKubeconfigSecretsRequest'
      responses:
        '200':
          description: List of kubeconfig secrets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      success:
                        allOf:
                          - $ref: '#/components/schemas/JSONSuccessResponse'
                          - type: object
                            properties:
                              code:
                                type: string
                                enum: [ok]
                              data:
                                $ref: '#/components/schemas/ListKubeconfigSecretsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
components:
  schemas:
    Response:
      type: object
      properties:
        success:
          $ref: '#/components/schemas/JSONSuccessResponse'
        error:
          $ref: '#/components/schemas/JSONErrorResponse'
    JSONSuccessResponse:
      type: object
      required:
        - code
        - data
      properties:
        code:
          type: string
          enum: [ok, created, deleted]
        data: {}
    JSONErrorResponse:
      type: object
      required:
        - code
        - data
      properties:
        code:
          type: string
          enum:
            - internal_error
            - not_found
            - bad_request
            - unsupported_media_type
            - malformed_json_body
            - body_validation
            - form_data_too_large
        data: {}
...
    KubeconfigSecretResponse:
      type: object
      required:
        - name
        - namespace
      properties:
        name:
          type: string
        namespace:
          type: string
    DeleteKubeconfigSecretRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$'
          maxLength: 253
          description: DNS 1123 subdomain
    ListKubeconfigSecretsRequest:
      type: object
      properties:
        limit:
          type: integer
          format: int64
          minimum: 1
          maximum: 250
          default: 25
        continueToken:
          type: string
    ListKubeconfigSecretsResponse:
      type: object
      required:
        - continueToken
        - secrets
      properties:
        continueToken:
          type: string
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/KubeconfigSecretResponse'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/JSONErrorResponse'
                      - type: object
                        properties:
                          code:
                            type: string
                            enum: [bad_request, malformed_json_body, body_validation]
    UnsupportedMediaType:
      description: Unsupported media type
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/JSONErrorResponse'
                      - type: object
                        properties:
                          code:
                            type: string
                            enum: [unsupported_media_type]
                          data:
                            type: string
    FormDataTooLarge:
      description: Form data too large
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Response'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/JSONErrorResponse'
                      - type: object
                        properties:
                          code:
                            type: string
                            enum: [form_data_too_large]
                          data:
                            type: string